Moved some visualization tools to utils
